2019-07-01: Added a new structure for pages
