+++ resources/js/mixins/Redirect.js
@@ -1,0 +1,7 @@
+export default {
+  methods: {
+    goBackToListPage() {
+      window.location.href = '/';
+    },
+  }
+};
